Added more instructions and spacing
Index.html: I made the instructions with a <p> tag so that I can access them andadd padding to make the spacing look better. There is also a line break on
line 22.
Styles.css: In footer of css, I made the display inline block and there
should be footer p with padding on ine 27.
I made these changes so that the spacing looks nicer and I don't have to
add so many line breaks.